requirements.txt
requests>=2.32.0
python-dotenv>=1.2.0
cryptography>=46.0.0
